+++ inc/staffSvg/index.ts
@@ -12,7 +12,7 @@ import LilyDocument from "../lilyParser/lilyDocument";
 import {LilyDocumentAttribute, LilyDocumentAttributeReadOnly} from "../lilyParser/lilyDocument";
 import {Scheme, SchemePair} from "../lilyParser/lilyTerms";
 import {SheetPage, SheetSystem, SheetStaff, SheetMeasure} from "./sheetDocument";
-import Glyph from "./glyph";
+import * as glyph from "./glyph";
 
 
 
@@ -178,5 +178,5 @@ export {
 	HashTable,
 	LilyDocumentAttribute,
 	LilyDocumentAttributeReadOnly,
-	Glyph,
+	glyph,
 };